python/pyproject.toml
[project]
name = "Extract-handwriting-from-substation-schematic"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Seth Deegan", email = ""},
]
dependencies = [
    "azure-cognitiveservices-vision-computervision>=0.9.0",
    "pillow>=10.3.0",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.0.1",
    "flask>=3.0.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
